+++ castor/lib/core/src/main/scala/io/iohk/atala/castor/core/model/did/w3c/package.scala
@@ -24,6 +24,7 @@ package object w3c {
           {
             case DIDResolutionError.DLTProxyError(_)       => DIDResolutionErrorRepr.InternalError
             case DIDResolutionError.UnexpectedDLTResult(_) => DIDResolutionErrorRepr.InternalError
+            case DIDResolutionError.ValidationError(_)     => DIDResolutionErrorRepr.InvalidDID
           },
           _.toRight(DIDResolutionErrorRepr.NotFound)
         )